refactor: cleaning codemirror code
